moved handle download
